cache: extract cache eviction heap
Start creating an interface that doesn't require using heap and hides more of the
entry internals.
